removing useless images

+++ Maintaining/UselessImagesFinder.java
@@ -16,7 +16,7 @@ public class UselessImagesFinder {
 	 *            the extension of the files
 	 * @return the {@link LinkedList} of the filenames, in {@link String}
 	 */
-	static LinkedList<String> files(String path, String extension) {
+	public static LinkedList<String> listFiles(String path, String extension) {
 		File folder = new File(path);
 		File[] listOfFiles = folder.listFiles();
 		LinkedList<String> rep = new LinkedList<String>();
@@ -40,8 +40,8 @@ public class UselessImagesFinder {
 	 * @param path
 	 *            the path to the images
 	 */
-	static void findUselessImages(ListOfWords list, String path) {
-		LinkedList<String> filenames = files(path, "jpg");
+	public static void findUselessImages(ListOfWords list, String path) {
+		LinkedList<String> filenames = listFiles(path, "jpg");
 		// for (String s : filenames)
 		// System.out.println(s);
 		System.out.println(" * Total number of images :" + filenames.size());
@@ -60,18 +60,31 @@ public class UselessImagesFinder {
 		for (String s : filenames) {
 			System.out.println(s);
 		}
-		
+
 		// compute the instruction
-		System.out.println("\n * Instruction :");
-		String instr = "rm ";
-		for (String s : filenames)
-			instr = instr + s + " ";
-		System.out.println(instr);
+		if (filenames.size() > 0) {
+			System.out.println("\n * Instruction :");
+			String instr = "rm ";
+			for (String s : filenames)
+				instr = instr + s + " ";
+			System.out.println(instr);
+		}
 	}
-	
-	static void findMissingPictures(ListOfWords list, String path) {
-		LinkedList<String> filenames = files(path, "jpg");
+
+	/**
+	 * display the {@link Word}s with a picture which is not available on hard
+	 * disk
+	 * 
+	 * @param list
+	 *            the {@link ListOfWords}
+	 * @param path
+	 *            the path to the pictures
+	 */
+	public static void findMissingPictures(ListOfWords list, String path) {
+		LinkedList<String> filenames = listFiles(path, "jpg");
 		System.out.println(" * Total number of images :" + filenames.size());
+
+		LinkedList<Word> missing = new LinkedList<Word>();
 
 		// remove all the used filenames
 		for (Word w : list.getWords()) {
@@ -79,16 +92,46 @@ public class UselessImagesFinder {
 				continue;
 			String pic_filename = w.getPictureFilename();
 			boolean contains = filenames.contains(pic_filename);
-			if (!contains) {
-				System.out.println("Missing :\n - " + pic_filename);
+			if (!contains)
+				missing.add(w);
+		}
+
+		System.out.println(" * Number of words with a missing pic : "
+				+ missing.size());
+
+		if (missing.size() > 0) {
+			for (Word w : missing) {
+				System.out.println("Missing :\n - " + w.getPictureFilename());
 				System.out.println(" - Word :" + w.toString_onlyWords());
 			}
 		}
 	}
 
+	/**
+	 * display the {@link Word}s with a picture which is not available on hard
+	 * disk
+	 * 
+	 * @param list
+	 *            the {@link ListOfWords}
+	 * @param path
+	 *            the path to the pictures
+	 */
+	public static void checkPictures(ListOfWords list,
+			String path) {
+		System.out.println("USELESS PICS");
+		findUselessImages(list, path);
+		System.out.println();
+		System.out.println("MISSING PICS");
+		findMissingPictures(list, path);
+	}
+
 	public static void main(String[] args) {
-		findUselessImages(ListOfWords.defaultListOfWords(),
+		 checkPictures(ListOfWords.defaultListOfWords(),
 				ImageGetter.DEFAULT_DIR);
+
+		// ListOfWords w = new ListOfWords();
+		// w.readFile("/test.kvtml");
+		//		checkPictures(w, ImageGetter.DEFAULT_DIR);
 	}
 
 }
